# swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"만세력 API","description":"사주팔자 분석 및 해석 API"}],"paths":{"/api/v1/manseryeok/points":{"post":{"tags":["만세력 API"],"summary":"오행/십성 분석 조회","description":"오행 분포와 십성 분석 정보를 조회합니다","operationId":"getOhaeng","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ManseryeokCreateRequest"}}},"required":true},"responses":{"400":{"description":"잘못된 요청","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OhaengCreateResponse"}}}},"200":{"description":"조회 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OhaengCreateResponse"}}}}}}},"/api/v1/manseryeok/interpretation":{"post":{"tags":["만세력 API"],"summary":"사주 해석 생성","description":"생년월일시 정보를 입력받아 AI 기반 사주팔자 해석을 제공합니다","operationId":"getInterpretation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ManseryeokCreateRequest"}}},"required":true},"responses":{"200":{"description":"해석 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ManseryeokInterpretationResponse"}}}},"400":{"description":"잘못된 요청","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ManseryeokInterpretationResponse"}}}},"500":{"description":"서버 오류","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ManseryeokInterpretationResponse"}}}}}}},"/api/v1/manseryeok/daeun":{"post":{"tags":["만세력 API"],"summary":"대운 정보 조회","description":"생년월일시를 바탕으로 대운, 연운, 월운 정보를 조회합니다","operationId":"getManseryeok","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ManseryeokCreateRequest"}}},"required":true},"responses":{"400":{"description":"잘못된 요청","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DaeunCreateResponse"}}}},"200":{"description":"조회 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DaeunCreateResponse"}}}}}}},"/api/v1/manseryeok/chart":{"post":{"tags":["만세력 API"],"summary":"사주 기본 차트 조회","description":"사주팔자 기본 구조(연주, 월주, 일주, 시주)를 조회합니다","operationId":"getChart","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ManseryeokCreateRequest"}}},"required":true},"responses":{"400":{"description":"잘못된 요청","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChartCreateResponse"}}}},"200":{"description":"조회 성공","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChartCreateResponse"}}}}}}}},"components":{"schemas":{"ManseryeokCreateRequest":{"type":"object","properties":{"name":{"type":"string"},"gender":{"type":"string"},"calendar":{"type":"string"},"birthday":{"type":"string"},"birthtime":{"type":"string"},"hmUnsure":{"type":"boolean"},"day":{"type":"integer","format":"int32"},"hour":{"type":"integer","format":"int32"},"locationId":{"type":"integer","format":"int32"},"midnightAdjust":{"type":"boolean"},"min":{"type":"integer","format":"int32"},"month":{"type":"integer","format":"int32"},"year":{"type":"integer","format":"int32"},"locationName":{"type":"string"}},"description":"사주 분석 요청 정보"},"AnalysisData":{"type":"object","properties":{"_오행":{"type":"array","items":{"$ref":"#/components/schemas/ElementInfo"}},"_십성":{"type":"array","items":{"$ref":"#/components/schemas/ElementInfo"}}}},"Element":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"chinese":{"type":"string"},"_음양":{"$ref":"#/components/schemas/EumYang"},"_오행":{"$ref":"#/components/schemas/Ohaeng"},"_십성":{"$ref":"#/components/schemas/Sipseong"}}},"ElementInfo":{"type":"object","properties":{"element":{"$ref":"#/components/schemas/Element"},"point":{"type":"number","format":"double"},"percent":{"type":"number","format":"double"},"description":{"type":"string"}}},"EumYang":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"chinese":{"type":"string"}}},"Ohaeng":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"chinese":{"type":"string"}}},"OhaengCreateResponse":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/AnalysisData"}}},"Sipseong":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"chinese":{"type":"string"}}},"ManseryeokInterpretationResponse":{"type":"object","properties":{"personalInfo":{"$ref":"#/components/schemas/PersonalInfo"},"interpretation":{"type":"string"}}},"PersonalInfo":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"gender":{"type":"string","enum":["MALE","FEMALE"]},"birthDate":{"type":"string"},"birthTime":{"type":"string"},"calendarType":{"type":"string","enum":["SOLAR","LUNAR"]},"city":{"type":"string"},"locationId":{"type":"string"},"midnightAdjust":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"},"timeUnknown":{"type":"boolean"}}},"DaeunCreateResponse":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/SajuData"}}},"DaeunInfo":{"type":"object","properties":{"age":{"type":"integer","format":"int32"},"year":{"type":"integer","format":"int32"},"month":{"type":"integer","format":"int32"},"_간지":{"$ref":"#/components/schemas/GanjiInfo"}}},"GanjiInfo":{"type":"object","properties":{"_천간":{"$ref":"#/components/schemas/Element"},"_지지":{"$ref":"#/components/schemas/Element"},"_지장간":{"type":"array","items":{"$ref":"#/components/schemas/JijangganInfo"}},"_운성":{"$ref":"#/components/schemas/UnseongInfo"}}},"JijangganInfo":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"chinese":{"type":"string"}}},"SajuData":{"type":"object","properties":{"_대운수":{"type":"integer","format":"int32"},"_대운간지":{"type":"string"},"_대운":{"type":"array","items":{"$ref":"#/components/schemas/DaeunInfo"}},"_연운":{"type":"array","items":{"$ref":"#/components/schemas/YeonunInfo"}},"_월운":{"type":"array","items":{"$ref":"#/components/schemas/WolunInfo"}}}},"UnseongInfo":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"chinese":{"type":"string"}}},"WolunInfo":{"type":"object","properties":{"age":{"type":"integer","format":"int32"},"year":{"type":"integer","format":"int32"},"month":{"type":"integer","format":"int32"},"_간지":{"$ref":"#/components/schemas/GanjiInfo"}}},"YeonunInfo":{"type":"object","properties":{"age":{"type":"integer","format":"int32"},"year":{"type":"integer","format":"int32"},"month":{"type":"integer","format":"int32"},"_간지":{"$ref":"#/components/schemas/GanjiInfo"}}},"BasicChartData":{"type":"object","properties":{"profile":{"$ref":"#/components/schemas/ProfileInfo"},"_기본명식":{"$ref":"#/components/schemas/SajuChart"},"_신살":{"$ref":"#/components/schemas/SinsalInfo"}}},"ChartCreateResponse":{"type":"object","properties":{"status":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/BasicChartData"}}},"Pillar":{"type":"object","properties":{"_천간":{"$ref":"#/components/schemas/Element"},"_지지":{"$ref":"#/components/schemas/Element"},"_지장간":{"type":"array","items":{"$ref":"#/components/schemas/JijangganInfo"}},"_운성":{"$ref":"#/components/schemas/UnseongInfo"}}},"ProfileInfo":{"type":"object","properties":{"index":{"type":"integer","format":"int32"},"avatar":{"type":"string"},"sexagenaryCycle":{"type":"string"},"sunBirth":{"type":"string"},"lunBirth":{"type":"string"},"adjustedBirth":{"type":"string"},"location":{"type":"string"},"adjusted":{"type":"string"}}},"SajuChart":{"type":"object","properties":{"_세차":{"$ref":"#/components/schemas/Pillar"},"_월건":{"$ref":"#/components/schemas/Pillar"},"_일진":{"$ref":"#/components/schemas/Pillar"},"_시진":{"$ref":"#/components/schemas/Pillar"}}},"SinsalElement":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"chinese":{"type":"string"}}},"SinsalInfo":{"type":"object","properties":{"_세차":{"$ref":"#/components/schemas/SinsalElement"},"_월건":{"$ref":"#/components/schemas/SinsalElement"},"_일진":{"$ref":"#/components/schemas/SinsalElement"},"_시진":{"$ref":"#/components/schemas/SinsalElement"}}}}}}